Needs a lot more testing and a lot more careful reading
This passes all the test, but it's probably not good enough.
There seems to be a lot of places where they call .instance_method when
they expect to get the method that is defined directly defined on a
class.

This is not good. Every instace when this happens it is a bug.